combined stop and cleanup in ContainerRunner

+++ save-orchestrator-common/src/main/kotlin/com/saveourtool/save/orchestrator/kubernetes/KubernetesManager.kt
@@ -93,18 +93,7 @@ class KubernetesManager(
         }
     }
 
-    override fun stop(executionId: Long) {
-        val jobName = jobNameForExecution(executionId)
-        val deletedResources = kcJobsWithName(jobName)
-            .delete()
-        val isDeleted = deletedResources.size == 1
-        if (!isDeleted) {
-            throw ContainerRunnerException("Failed to delete job with name $jobName: response is $deletedResources")
-        }
-        logger.debug("Deleted Job for execution id=$executionId")
-    }
-
-    override fun stopByContainerId(containerId: String): Boolean {
+    override fun stop(containerId: String): Boolean {
         logger.warn {
             "${this::class.simpleName}#stopByAgentId is called, but it's no-op, " +
                     "because we don't directly delete pods in kubernetes"
@@ -112,12 +101,19 @@ class KubernetesManager(
         return false
     }
 
-    override fun cleanup(executionId: Long) {
-        logger.debug("Removing a Job for execution id=$executionId")
-        val job = kcJobsWithName(jobNameForExecution(executionId))
-        job.get()?.let {
-            job.delete()
+    override fun cleanupByExecution(executionId: Long) {
+        logger.debug("Removing a job for execution id=$executionId")
+        val jobName = jobNameForExecution(executionId)
+        val job = kcJobsWithName(jobName)
+        if (job.get() == null) {
+            logger.warn { "Failed to delete job with name $jobName: there is no such job" }
+            return
         }
+        val deletedResources = job.delete()
+        if (deletedResources.size != 1) {
+            throw ContainerRunnerException("Failed to delete job with name $jobName: response is $deletedResources")
+        }
+        logger.debug("Cleanup job for execution id=$executionId")
     }
 
     override fun prune() {
@@ -131,7 +127,7 @@ class KubernetesManager(
         .items
         .map { it.metadata.name }
 
-    override fun isStoppedByContainerId(containerId: String): Boolean {
+    override fun isStopped(containerId: String): Boolean {
         val pod = kc.pods().withName(containerId).get()
         return pod == null || run {
             // Retrieve reason based on https://github.com/kubernetes/kubernetes/issues/22839
